Move loss utils to common policy utils
