+++ library/src/physicalai/policies/pi05/model.py
@@ -19,9 +19,10 @@ from transformers.cache_utils import DynamicCache
 
 from physicalai.data.constants import IMAGE_MASKS, TOKENIZED_PROMPT, TOKENIZED_PROMPT_MASK
 from physicalai.data.observation import ACTION, IMAGES
-from physicalai.policies.base import Model, in_episode_bound, reduce_losses
+from physicalai.policies.base import Model
 from physicalai.policies.mixins import SnapFlowModelMixin
 from physicalai.policies.mixins.peft import PeftModelMixin
+from physicalai.policies.utils import in_episode_bound, reduce_losses
 
 from .pi_gemma import (
     PaliGemmaForConditionalGenerationWithPiGemma,
